+++ fitcheck/views/review.py
@@ -11,7 +11,8 @@ from django.views.decorators.http import require_POST
 from ..forms import ReviewDecisionForm
 from ..models import Doctrine, FitSubmission
 from ..services.check_runner import review_submission
-from .member import _can_review, _paginate
+from .common import paginate as _paginate
+from .member import _can_review
 
 
 def review_access_required(view):